# [SPARK-36444][SQL][3.1] Remove OptimizeSubqueries from batch of PartitionPruning