### What changes were proposed in this pull request?

This is a backport PR of #33664 for branch-3.1.

### Why are the changes needed?

We found a query in production that cost lots of time in optimize phase when enable DPP, the SQL pattern likes

```
select <cols...>
from a
left join b on a.<col> = b.<col>
left join c on b.<col> = c.<col>
left join d on c.<col> = d.<col>
left join e on d.<col> = e.<col>
left join f on e.<col> = f.<col>
left join g on f.<col> = g.<col>
left join h on g.<col> = h.<col>
...
```

<details>

<summary>Before this PR, Analyzer/Optimizer costs 15821.6 seconds</summary>

```
=== Metrics of Analyzer/Optimizer Rules ===
Total number of runs: 24763671
Total time: 15821.588159083 seconds

Rule                                                                                               Effective Time / Total Time                     Effective Runs / Total Runs

org.apache.spark.sql.catalyst.optimizer.Optimizer$OptimizeSubqueries                               6325784430499 / 14213630904736                  2047 / 342802
org.apache.spark.sql.catalyst.optimizer.ColumnPruning                                              28009746 / 267164777817                         1 / 691736
org.apache.spark.sql.catalyst.analysis.UpdateAttributeNullability                                  0 / 118329246545                                0 / 342804
org.apache.spark.sql.catalyst.optimizer.RemoveRedundantAliases                                     8997600 / 71070686823                           1 / 348934
org.apache.spark.sql.catalyst.optimizer.BooleanSimplification                                      923149600 / 59903945393                         2036 / 348934
org.apache.spark.sql.catalyst.optimizer.NullPropagation                                            12343892 / 52895299683                          1 / 348934
org.apache.spark.sql.execution.datasources.SchemaPruning                                           0 / 51638842573                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.UnwrapCastInBinaryComparison                               0 / 42686293006                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.SimplifyConditionals                                       0 / 41190229736                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.InferFiltersFromConstraints                                1482075440 / 40899690529                        2048 / 171401
org.apache.spark.sql.catalyst.optimizer.SimplifyBinaryComparison                                   0 / 40747831162                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.SimplifyCaseConversionExpressions                          0 / 39946334062                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.FoldablePropagation                                        655084587 / 39426073462                         2047 / 348934
org.apache.spark.sql.catalyst.optimizer.PruneFilters                                               489780 / 34731670364                            1 / 520335
org.apache.spark.sql.catalyst.optimizer.OptimizeJsonExprs                                          0 / 33533545307                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.ReplaceNullWithFalseInPredicate                            0 / 31910754610                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.ConstantFolding                                            22871375 / 31308048659                          1 / 348934
org.apache.spark.sql.catalyst.optimizer.LikeSimplification                                         0 / 31276042884                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.OptimizeWindowFunctions                                    0 / 31100888868                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.ReorderAssociativeOperator                                 0 / 30855866688                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.SimplifyCasts                                              0 / 30414758547                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.OptimizeIn                                                 0 / 30211101358                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.OptimizeUpdateFields                                       0 / 29986491532                                 0 / 348936
org.apache.spark.sql.catalyst.optimizer.RemoveDispensableExpressions                               0 / 29740191900                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.SimplifyExtractValueOps                                    0 / 25399289000                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.RewriteCorrelatedScalarSubquery                            0 / 21381558331                                 0 / 348934
org.apache.spark.sql.catalyst.optimizer.PushDownPredicates                                         5831922974 / 18327095753                        3072 / 691737
org.apache.spark.sql.catalyst.optimizer.ReplaceExpressions                                         5007709 / 17431480903                           1 / 171401
org.apache.spark.sql.catalyst.optimizer.GetCurrentDatabaseAndCatalog                               0 / 15321596322                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.ComputeCurrentTime                                         0 / 15060720651                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.DecimalAggregates                                          0 / 14874103589                                 0 / 171401
org.apache.spark.sql.execution.datasources.PruneFileSourcePartitions                               50420951 / 14798400219                          1 / 171401
org.apache.spark.sql.catalyst.optimizer.ReplaceUpdateFieldsExpression                              0 / 14791439186                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.EliminateMapObjects                                        0 / 14361213979                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.ReassignLambdaVariableID                                   0 / 14258459712                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.RewriteNonCorrelatedExists                                 0 / 14094898098                                 0 / 171401
org.apache.spark.sql.catalyst.optimizer.RemoveNoopOperators                                        1975195 / 12967136767                           2 / 691736
org.apache.spark.sql.catalyst.optimizer.PullupCorrelatedPredicates                                 0 / 12573667353                                 0 / 171401
org.apache.spark.sql.execution.datasources.v2.V2ScanRelationPushDown                               0 / 11355809540                                 0 / 171401
org.apache.spark.sql.execution.dynamicpruning.CleanupDynamicPruningFilters                         350911193 / 9931582633                          2048 / 171401
org.apache.spark.sql.catalyst.optimizer.ConstantPropagation                                        0 / 8268511133                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.CollapseProject                                            10135014 / 6764166524                           1 / 520335
org.apache.spark.sql.catalyst.optimizer.EliminateResolvedHint                                      0 / 6244880486                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.CombineUnions                                              0 / 5811583863                                  0 / 520335
org.apache.spark.sql.catalyst.optimizer.RewriteDistinctAggregates                                  28926471 / 5423624109                           1 / 171401
org.apache.spark.sql.catalyst.optimizer.CollapseRepartition                                        0 / 5245332974                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.EliminateSorts                                             0 / 5051398253                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.PropagateEmptyRelation                                     0 / 4975057807                                  0 / 342802
org.apache.spark.sql.catalyst.optimizer.NormalizeFloatingNumbers                                   0 / 4804919393                                  0 / 171401
org.apache.spark.sql.execution.python.ExtractPythonUDFFromAggregate                                0 / 4408441952                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.CollapseWindow                                             0 / 4349408795                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.EliminateSerialization                                     0 / 4170274724                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.TransposeWindow                                            0 / 4113327220                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.ReorderJoin                                                0 / 4107443950                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.CombineFilters                                             0 / 4061601886                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.PushProjectionThroughUnion                                 0 / 4042615936                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.PushDownLeftSemiAntiJoin                                   0 / 4024293929                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.EliminateLimits                                            0 / 3908219408                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.RewritePredicateSubquery                                   135506567 / 3895045754                          2047 / 171401
org.apache.spark.sql.catalyst.optimizer.EliminateOuterJoin                                         0 / 3875669986                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.PushLeftSemiLeftAntiThroughJoin                            0 / 3875663441                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.LimitPushDown                                              0 / 3859481346                                  0 / 348934
org.apache.spark.sql.catalyst.optimizer.ReplaceDeduplicateWithAggregate                            0 / 3404517096                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.PushExtraPredicateThroughJoin                              0 / 2876802099                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.PushPredicateThroughNonJoin                                0 / 2546037832                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ExtractPythonUDFFromJoinCondition                          0 / 2392444200                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.RemoveLiteralFromGroupExpressions                          0 / 2297492778                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.RewriteExceptAll                                           0 / 2239751163                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.InferFiltersFromGenerate                                   0 / 2189398940                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.RemoveRepetitionFromGroupExpressions                       0 / 2187645769                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ReplaceExceptWithFilter                                    0 / 2135169574                                  0 / 171401
org.apache.spark.sql.execution.python.ExtractGroupingPythonUDFFromAggregate                        0 / 2058019085                                  0 / 171401
org.apache.spark.sql.catalyst.analysis.EliminateSubqueryAliases                                    602948 / 2030110600                             1 / 171401
org.apache.spark.sql.catalyst.optimizer.OptimizeLimitZero                                          0 / 2028208284                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.CombineTypedFilters                                        0 / 1985665817                                  0 / 171401
org.apache.spark.sql.catalyst.analysis.EliminateView                                               0 / 1965869470                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ObjectSerializerPruning                                    0 / 1892132711                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.RewriteIntersectAll                                        0 / 1891839429                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ReplaceIntersectWithSemiJoin                               0 / 1873028976                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ReplaceDistinctWithAggregate                               0 / 1872377392                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.ReplaceExceptWithAntiJoin                                  0 / 1860128741                                  0 / 171401
org.apache.spark.sql.catalyst.optimizer.CheckCartesianProducts                                     0 / 405470117                                   0 / 342802
org.apache.spark.sql.execution.OptimizeMetadataOnlyQuery                                           0 / 242854215                                   0 / 171401
org.apache.spark.sql.catalyst.optimizer.CostBasedJoinReorder                                       0 / 226833963                                   0 / 171401
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveReferences                                  117151873 / 123533385                           7 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$FunctionArgumentConversion                     102066025 / 121689146                           5 / 13
org.apache.spark.sql.catalyst.optimizer.CombineConcats                                             0 / 112256648                                   0 / 348934
org.apache.spark.sql.catalyst.analysis.TypeCoercion$ImplicitTypeCasts                              76183548 / 108546928                            4 / 13
org.apache.spark.sql.catalyst.optimizer.EliminateAggregateFilter                                   0 / 107905167                                   0 / 348934
org.apache.spark.sql.catalyst.optimizer.EliminateDistinct                                          0 / 103804760                                   0 / 171401
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveFunctions                                   85607389 / 94162213                             8 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$Division                                       76446295 / 93129469                             5 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$PromoteStrings                                 53450988 / 87954605                             1 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ExtractGenerator                                   0 / 85268434                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$IfCoercion                                     50348094 / 80404327                             4 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$CaseWhenCoercion                               39786113 / 59089899                             3 / 13
org.apache.spark.sql.catalyst.analysis.ResolveTimeZone                                             56110133 / 57619849                             10 / 13
org.apache.spark.sql.execution.dynamicpruning.PartitionPruning                                     22888739 / 54329716                             1 / 171401
org.apache.spark.sql.catalyst.analysis.DecimalPrecision                                            0 / 49832160                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$InConversion                                   0 / 46158859                                    0 / 13
org.apache.spark.sql.execution.datasources.FindDataSourceTable                                     43323608 / 44575170                             1 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$ConcatCoercion                                 14900959 / 39466022                             1 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveBinaryArithmetic                            0 / 38815701                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$BooleanEquality                                0 / 36470552                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$DateTimeOperations                             7216638 / 35530574                              1 / 13
org.apache.spark.sql.execution.python.ExtractPythonUDFs                                            0 / 34271447                                    0 / 171401
org.apache.spark.sql.catalyst.analysis.ResolveLambdaVariables                                      0 / 32166544                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$IntegralDivision                               0 / 30547791                                    0 / 13
org.apache.spark.sql.catalyst.analysis.ResolveExpressionsWithNamePlaceholders                      0 / 30506602                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$StringLiteralCoercion                          0 / 30116261                                    0 / 13
org.apache.spark.sql.catalyst.analysis.ResolveHigherOrderFunctions                                 0 / 27859446                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$EltCoercion                                    0 / 24218064                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$WindowFrameCoercion                            0 / 23413444                                    0 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveRandomSeed                                  0 / 22036715                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$StackCoercion                                  0 / 21957835                                    0 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveWindowOrder                                 0 / 21463552                                    0 / 13
org.apache.spark.sql.catalyst.analysis.TypeCoercion$MapZipWithCoercion                             0 / 20489026                                    0 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveWindowFrame                                 0 / 20481142                                    0 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$GlobalAggregates                                   0 / 19715779                                    0 / 13
org.apache.spark.sql.catalyst.analysis.Analyzer$ResolveMissingReferences                           0 / 19693693                                    0 / 13
…