@pan3793 2022-02-08
> Apache Spark community follows Semantic Versioning and doesn't allow backporting of new feature or improvements.

Thanks @dongjoon-hyun I think maybe we can treat it as a bug based on https://github.com/apache/spark/pull/33664#issuecomment-901589931, comparing to 2.4s Analyzer/Optimizer costs 15821.6 seconds looks really unreasonable.

> To be clear, I'm asking you to upgrade Kyuubi to 3.2.1 instead of 3.1.2